move getMaxSequenceLimit from ChangeFilesUtil to TableOptimizeItem
