+++ ams/ams-server/src/main/java/com/netease/arctic/ams/server/optimize/TableOptimizeItem.java
@@ -44,7 +44,6 @@ import com.netease.arctic.ams.server.model.TableTaskHistory;
 import com.netease.arctic.ams.server.service.IJDBCService;
 import com.netease.arctic.ams.server.service.IQuotaService;
 import com.netease.arctic.ams.server.service.ServiceContainer;
-import com.netease.arctic.ams.server.utils.ChangeFilesUtil;
 import com.netease.arctic.ams.server.utils.FilesStatisticsBuilder;
 import com.netease.arctic.ams.server.utils.TableStatCollector;
 import com.netease.arctic.ams.server.utils.UnKeyedTableUtil;
@@ -1126,15 +1125,6 @@ public class TableOptimizeItem extends IJDBCService {
         baseSnapshot == null ? TableOptimizeRuntime.INVALID_SNAPSHOT_ID : baseSnapshot.snapshotId());
   }
 
-  /**
-   * Select all the files whose sequence <= maxSequence as Selected-Files, seek the maxSequence to find as many
-   * Selected-Files as possible, and also
-   * - the cnt of these Selected-Files must <= maxFileCntLimit
-   * - the max TransactionId of the Selected-Files must > the min TransactionId of all the left files
-   *
-   * @return the max sequence of selected file, return Long.MAX_VALUE if all files should be selected,
-   * Long.MIN_VALUE means no files should be selected
-   */
   private long getMaxSequenceLimit(Snapshot changeSnapshot,
                                    StructLikeMap<Long> partitionOptimizedSequence,
                                    StructLikeMap<Long> legacyPartitionMaxTransactionId) {
@@ -1152,13 +1142,13 @@ public class TableOptimizeItem extends IJDBCService {
             .fromSequence(partitionOptimizedSequence)
             .fromLegacyTransaction(legacyPartitionMaxTransactionId)
             .useSnapshot(changeSnapshot.snapshotId());
-    Map<Long, ChangeFilesUtil.SnapshotFileGroup> changeFilesGroupBySequence = new HashMap<>();
+    Map<Long, SnapshotFileGroup> changeFilesGroupBySequence = new HashMap<>();
     try (CloseableIterable<ContentFileWithSequence<?>> files = changeTableIncrementalScan.planFilesWithSequence()) {
       for (ContentFileWithSequence<?> file : files) {
-        ChangeFilesUtil.SnapshotFileGroup fileGroup =
+        SnapshotFileGroup fileGroup =
             changeFilesGroupBySequence.computeIfAbsent(file.getSequenceNumber(), key -> {
               long txId = FileNameGenerator.parseChangeTransactionId(file.path().toString(), file.getSequenceNumber());
-              return new ChangeFilesUtil.SnapshotFileGroup(file.getSequenceNumber(), txId);
+              return new SnapshotFileGroup(file.getSequenceNumber(), txId);
             });
         fileGroup.addFile();
       }
@@ -1173,7 +1163,7 @@ public class TableOptimizeItem extends IJDBCService {
     }
 
     long maxSequence =
-        ChangeFilesUtil.getMaxSequenceLimit(new ArrayList<>(changeFilesGroupBySequence.values()), maxFileCntLimit);
+        getMaxSequenceKeepingTxIdInOrder(new ArrayList<>(changeFilesGroupBySequence.values()), maxFileCntLimit);
     if (maxSequence == Long.MIN_VALUE) {
       LOG.warn("{} get no change files with self-optimizing.max-file-count={}, change it to a bigger value",
           tableIdentifier, maxFileCntLimit);
@@ -1182,6 +1172,84 @@ public class TableOptimizeItem extends IJDBCService {
           tableIdentifier, maxFileCntLimit, maxSequence);
     }
     return maxSequence;
+  }
+
+  /**
+   * Select all the files whose sequence <= maxSequence as Selected-Files, seek the maxSequence to find as many
+   * Selected-Files as possible, and also
+   * - the cnt of these Selected-Files must <= maxFileCntLimit
+   * - the max TransactionId of the Selected-Files must > the min TransactionId of all the left files
+   *
+   * @param snapshotFileGroups snapshotFileGroups
+   * @param maxFileCntLimit    maxFileCntLimit
+   * @return the max sequence of selected file, return Long.MAX_VALUE if all files should be selected,
+   * Long.MIN_VALUE means no files should be selected
+   */
+  static long getMaxSequenceKeepingTxIdInOrder(List<SnapshotFileGroup> snapshotFileGroups, long maxFileCntLimit) {
+    if (maxFileCntLimit <= 0 || snapshotFileGroups == null || snapshotFileGroups.isEmpty()) {
+      return Long.MIN_VALUE;
+    }
+    // 1.sort sequence
+    Collections.sort(snapshotFileGroups);
+    // 2.find the max index where all file cnt <= maxFileCntLimit
+    int index = -1;
+    int fileCnt = 0;
+    for (int i = 0; i < snapshotFileGroups.size(); i++) {
+      fileCnt += snapshotFileGroups.get(i).getFileCnt();
+      if (fileCnt <= maxFileCntLimit) {
+        index = i;
+      } else {
+        break;
+      }
+    }
+    // all files cnt <= maxFileCntLimit, return all files
+    if (fileCnt <= maxFileCntLimit) {
+      return Long.MAX_VALUE;
+    }
+    // if only check the first file groups, then the file cnt > maxFileCntLimit, no files should be selected
+    if (index == -1) {
+      return Long.MIN_VALUE;
+    }
+
+    // 3.wrap file group with the max TransactionId before and min TransactionId after
+    List<SnapshotFileGroupWrapper> snapshotFileGroupWrappers = wrapMinMaxTransactionId(snapshotFileGroups);
+    // 4.find the valid snapshotFileGroup
+    while (true) {
+      SnapshotFileGroupWrapper current = snapshotFileGroupWrappers.get(index);
+      // check transaction id inorder: max transaction id before(inclusive) < min transaction id after
+      if (Math.max(current.getFileGroup().getTransactionId(), current.getMaxTransactionIdBefore()) <
+          current.getMinTransactionIdAfter()) {
+        return current.getFileGroup().getSequence();
+      }
+      index--;
+      if (index == -1) {
+        return Long.MIN_VALUE;
+      }
+    }
+  }
+
+  private static List<SnapshotFileGroupWrapper> wrapMinMaxTransactionId(List<SnapshotFileGroup> snapshotFileGroups) {
+    List<SnapshotFileGroupWrapper> wrappedList = new ArrayList<>();
+    for (SnapshotFileGroup snapshotFileGroup : snapshotFileGroups) {
+      wrappedList.add(new SnapshotFileGroupWrapper(snapshotFileGroup));
+    }
+    long maxValue = Long.MIN_VALUE;
+    for (int i = 0; i < wrappedList.size(); i++) {
+      SnapshotFileGroupWrapper wrapper = wrappedList.get(i);
+      wrapper.setMaxTransactionIdBefore(maxValue);
+      if (wrapper.getFileGroup().getTransactionId() > maxValue) {
+        maxValue = wrapper.getFileGroup().getTransactionId();
+      }
+    }
+    long minValue = Long.MAX_VALUE;
+    for (int i = wrappedList.size() - 1; i >= 0; i--) {
+      SnapshotFileGroupWrapper wrapper = wrappedList.get(i);
+      wrapper.setMinTransactionIdAfter(minValue);
+      if (wrapper.getFileGroup().getTransactionId() < minValue) {
+        minValue = wrapper.getFileGroup().getTransactionId();
+      }
+    }
+    return wrappedList;
   }
 
   /**
@@ -1274,5 +1342,81 @@ public class TableOptimizeItem extends IJDBCService {
     }
     OptimizeTaskItem optimizeTaskItem = new ArrayList<>(optimizeTasks.values()).get(0);
     return optimizeTaskItem.getTaskId().getType() == OptimizeType.Minor;
+  }
+
+  /**
+   * Files grouped by snapshot, but only with the file cnt.
+   */
+  static class SnapshotFileGroup implements Comparable<SnapshotFileGroup> {
+    private final long sequence;
+    private final long transactionId;
+    private int fileCnt = 0;
+
+    public SnapshotFileGroup(long sequence, long transactionId) {
+      this.sequence = sequence;
+      this.transactionId = transactionId;
+    }
+
+    public SnapshotFileGroup(long sequence, long transactionId, int fileCnt) {
+      this.sequence = sequence;
+      this.transactionId = transactionId;
+      this.fileCnt = fileCnt;
+    }
+
+    public void addFile() {
+      fileCnt++;
+    }
+
+    public long getTransactionId() {
+      return transactionId;
+    }
+
+    public int getFileCnt() {
+      return fileCnt;
+    }
+
+    public long getSequence() {
+      return sequence;
+    }
+
+    @Override
+    public int compareTo(SnapshotFileGroup o) {
+      return Long.compare(this.sequence, o.sequence);
+    }
+  }
+
+  /**
+   * Wrap SnapshotFileGroup with max and min Transaction Id
+   */
+  private static class SnapshotFileGroupWrapper {
+    private final SnapshotFileGroup fileGroup;
+    // in the ordered file group list, the max transaction before this file group, Long.MIN_VALUE for the first
+    private long maxTransactionIdBefore;
+    // in the ordered file group list, the min transaction after this file group, Long.MAX_VALUE for the last
+    private long minTransactionIdAfter;
+
+    public SnapshotFileGroupWrapper(SnapshotFileGroup fileGroup) {
+      this.fileGroup = fileGroup;
+    }
+
+    public SnapshotFileGroup getFileGroup() {
+      return fileGroup;
+    }
+
+    public long getMaxTransactionIdBefore() {
+      return maxTransactionIdBefore;
+    }
+
+    public void setMaxTransactionIdBefore(long maxTransactionIdBefore) {
+      this.maxTransactionIdBefore = maxTransactionIdBefore;
+    }
+
+    public long getMinTransactionIdAfter() {
+      return minTransactionIdAfter;
+    }
+
+    public void setMinTransactionIdAfter(long minTransactionIdAfter) {
+      this.minTransactionIdAfter = minTransactionIdAfter;
+    }
   }
 }
