+++ ams/ams-server/src/test/java/com/netease/arctic/ams/server/AmsTestBase.java
@@ -31,6 +31,7 @@ import com.netease.arctic.ams.server.handler.impl.ArcticTableMetastoreHandler;
 import com.netease.arctic.ams.server.handler.impl.OptimizeManagerHandler;
 import com.netease.arctic.ams.server.optimize.OptimizeService;
 import com.netease.arctic.ams.server.optimize.SupportHiveTestGroup;
+import com.netease.arctic.ams.server.optimize.TableOptimizeItemTest;
 import com.netease.arctic.ams.server.optimize.TestExpireFileCleanSupportIceberg;
 import com.netease.arctic.ams.server.optimize.TestExpiredFileClean;
 import com.netease.arctic.ams.server.optimize.TestIcebergFullOptimizeCommit;
@@ -63,7 +64,6 @@ import com.netease.arctic.ams.server.service.impl.TableBlockerService;
 import com.netease.arctic.ams.server.service.impl.TestTableBlockerService;
 import com.netease.arctic.ams.server.util.DerbyTestUtil;
 import com.netease.arctic.ams.server.utils.CatalogUtil;
-import com.netease.arctic.ams.server.utils.ChangeFilesUtilTest;
 import com.netease.arctic.ams.server.utils.JDBCSqlSessionFactoryProvider;
 import com.netease.arctic.ams.server.utils.SequenceNumberFetcherTest;
 import com.netease.arctic.ams.server.utils.UnKeyedTableUtilTest;
@@ -127,7 +127,9 @@ import static org.powermock.api.mockito.PowerMockito.when;
     SupportHiveTestGroup.class,
     TestArcticTransactionService.class,
     TestOptimizerService.class,
-    UnKeyedTableUtilTest.class
+    UnKeyedTableUtilTest.class,
+    SequenceNumberFetcherTest.class,
+    TableOptimizeItemTest.class
 })
 @PrepareForTest({
     JDBCSqlSessionFactoryProvider.class,
@@ -143,9 +145,7 @@ import static org.powermock.api.mockito.PowerMockito.when;
     OptimizeManagerHandler.class,
     AdaptHiveService.class,
     PlatformFileInfoService.class,
-    HiveTableUtil.class,
-    SequenceNumberFetcherTest.class,
-    ChangeFilesUtilTest.class
+    HiveTableUtil.class
 })
 @PowerMockIgnore({"javax.management.*", "javax.net.ssl.*"})
 public class AmsTestBase {
